refactor(speclayout): rename designSpecs.js as specLayout.js
make the file name more sense: color and font parameters are  also design specs, but they are
defined in other files.
