+++ src/blocks/ImageLink.js
@@ -2,7 +2,7 @@ import styled from 'styled-components';
 import PropTypes from 'prop-types';
 
 import Img from 'src/elements/Img';
-import {header} from 'src/utils/designSpecs';
+import {header} from 'src/utils/specLayout';
 import {mediaQuery} from 'src/utils/breakpoints';
 import remify from 'src/utils/remify';
