+++ src/blocks/Main.js
@@ -7,7 +7,7 @@ import {
   header,
   page,
   twoColumns,
-} from 'src/utils/designSpecs';
+} from 'src/utils/specLayout';
 import remify from 'src/utils/remify';
 import {scale} from 'src/utils/specFont';
