+++ src/blocks/SectionColorData.js
@@ -1,6 +1,6 @@
 import styled from 'styled-components';
 
-import {formNumberLarge} from 'src/utils/designSpecs';
+import {formNumberLarge} from 'src/utils/specLayout';
 import {mediaQuery} from 'src/utils/breakpoints';
 import remify from 'src/utils/remify';
 import {scale} from 'src/utils/specFont';
