+++ src/blocks/WrapperColorData.js
@@ -2,7 +2,7 @@ import styled from 'styled-components';
 
 import {mediaQuery} from 'src/utils/breakpoints';
 import {boxSize} from 'src/utils/layout';
-import {formNumberLarge} from 'src/utils/designSpecs';
+import {formNumberLarge} from 'src/utils/specLayout';
 import remify from 'src/utils/remify';
 import {scale} from 'src/utils/specFont';
